Add config option to restore submission closure E-mail to curator
The new site-wide user account options have changed the default
behaviour of sending a copy of the submission closure message to the
curator. Curators can set their preferred behaviour if the system is
using site-wide user accounts, but this has changed the behaviour if
only local accounts are being used. For this case, there is now an
option 'submission_email_cc' that can be set in bigsdb.conf.