Microreact: Fix to support API changes.
Note that an API token now needs to be defined in the bigsdb.conf file.

This version also no longer requires fields to be named 'country' or
'year'. If you have a country field that is not named 'country' then
this can be set in the config.xml file using the
'microreact_country_field' attribute. If you have a year field that is
not named 'year' then this can be set in the config.xml file using the
'microreact_year_field' attribute.

Fixes #813.