Run plugins with the admin curator interface dataset if appropriate
Make sure that the datasets available to the plugin tools when run as
offline jobs are the same as those available in the interface used to
run the job. Admins in the curator interface can see private data
belonging to others but not when run from the standard user interface.

Fixes #849.