Restrict which isolates can be added/removed from project.
Don't allow project users to remove or add records from other users or
groups if restrict_user or restrict_usergroup is set on the project
unless they are an admin. Previously this only affected adding isolates
via a query, but for large projects like EMERT-II it would have been
possible for any member of the project to remove anybody else's isolates
from the project by modifying the isolate list.

Closes #931.