Remove PluginManager isolate count check.
This was only used to decide whether to list a plugin on the contents
page, but most plugins now are listed irrespective of the size of the
database since they allow lists of isolates to be entered. On large
databases with complex user views (due to private data etc.), this check
could take about 1 second which impacted page load time for negligible
benefit.