Prevent "SSL error: decryption failed or bad record mac"
If the process forks and then makes a database call you will get this
error. The database call is needed for the TreeViewer javascript, but
this is not needed if the scan has started, so check for this and do not
load the TreeViewer JS if we're not showing a tree.