Fix queries where a very long field name is used for seq attribs.
There is a limit to table name length in Pg of 63 characters. If a temp
table is needed for a query a temp table was created that combined the
locus and field names, but the table name could be too long. Now, if
this happens, a table is created that uses an MD5 hash of the locus and
field names. A temp file is now created that can be used to look up what
the locus and field are from the table name.